% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seqinfo.R
\name{cal_sc}
\alias{cal_sc}
\title{Calculate Sequence Complexity of a sequence}
\usage{
cal_sc(sequence, max_k = NULL)
}
\arguments{
\item{sequence}{The input DNA sequence.}

\item{max_k}{The maximum k value for kmers.}
}
\value{
The Sequence Complexity value.
}
\description{
Calculate Sequence Complexity of a sequence
}
\references{
https://resources.qiagenbioinformatics.com/manuals/clccancerresearchworkbench/200/index.php?manual=How_sequence_complexity_is_calculated.html
}
