% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seqinfo.R
\name{cal_shannon}
\alias{cal_shannon}
\title{Calculate Shannon Complexity of a sequence}
\usage{
cal_shannon(sequence, base = NULL)
}
\arguments{
\item{sequence}{The input DNA sequence.}

\item{base}{default the number of elements type in sequence.}
}
\value{
The Shannon Complexity value.
}
\description{
Calculate Shannon Complexity of a sequence
}
\references{
\enumerate{
\item Konopka, A. K. Sequence Complexity and Composition. in eLS (ed. John Wiley & Sons, Ltd) (Wiley, 2005). doi:10.1038/npg.els.0005260.
}
}
