% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_structrue.R
\name{calculate_identity}
\alias{calculate_identity}
\title{Calculate Sequence Identity of two sequence}
\usage{
calculate_identity(s1, s2, xmode = FALSE)
}
\arguments{
\item{s1}{First sequence}

\item{s2}{Second sequence}

\item{xmode}{consider the reverse string (all), complement and reverse complement string (DNA)}
}
\value{
Sequence identity as a decimal value between 0 and 1
}
\description{
This function calculates the sequence identity between two sequences.
}
\examples{
calculate_identity("ATCGTACG","ATCGTAGC")
calculate_identity("ATCGTACG","ATCGTAGC",xmode=T)
}
