% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{combine.crispr}
\alias{combine.crispr}
\title{Combine some crispr or multi_crispr object}
\usage{
\method{combine}{crispr}(...)
}
\arguments{
\item{...}{crispr or multi_crispr object}
}
\value{
multi_crispr object
}
\description{
Combine some crispr or multi_crispr object
}
