% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_type.R
\name{combine.evidence_level}
\alias{combine.evidence_level}
\title{Combine some evidence_level object}
\usage{
\method{combine}{evidence_level}(...)
}
\arguments{
\item{...}{evidence_level object}
}
\value{
evidence_level object
}
\description{
Combine some evidence_level object
}
