% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{combine}
\alias{combine}
\title{Combine some kinds of objects in \code{iCRISPR}}
\usage{
combine(...)
}
\arguments{
\item{...}{additional}
}
\value{
same object
}
\description{
Combine some kinds of objects in \code{iCRISPR}
}
