% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{combine.s_t_count}
\alias{combine.s_t_count}
\title{Combine some s_t_count object}
\usage{
\method{combine}{s_t_count}(...)
}
\arguments{
\item{...}{s_t_count object}
}
\value{
s_t_count object
}
\description{
Combine some s_t_count object
}
