% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{combine.s_t_gng}
\alias{combine.s_t_gng}
\title{Combine some s_t_gng object}
\usage{
\method{combine}{s_t_gng}(...)
}
\arguments{
\item{...}{s_t_gng object}
}
\value{
s_t_gng object
}
\description{
Combine some s_t_gng object
}
