% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{combine.s_t_pd}
\alias{combine.s_t_pd}
\title{Combine some s_t_pd object}
\usage{
\method{combine}{s_t_pd}(...)
}
\arguments{
\item{...}{s_t_pd object}
}
\value{
s_t_pd object
}
\description{
Combine some s_t_pd object
}
