% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{combine.self_df}
\alias{combine.self_df}
\title{Combine some self_df object}
\usage{
\method{combine}{self_df}(...)
}
\arguments{
\item{...}{self_df object}
}
\value{
self_df object
}
\description{
Combine some self_df object
}
