% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seqinfo.R
\name{count_kmers}
\alias{count_kmers}
\title{Count kmers in a sequence}
\usage{
count_kmers(sequence, k_size)
}
\arguments{
\item{sequence}{The input DNA sequence.}

\item{k_size}{The length of the kmers.}
}
\value{
A named vector with kmer counts.
}
\description{
Count kmers in a sequence
}
