% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{crispr}
\alias{crispr}
\title{One crispr object result from MAG_test.}
\format{
contains CRISPR, Cas, Arrary and Spacer
\describe{
\item{CRISPR}{CRISPR found in the whole genome (with multi-contigs), CRISPR_id is the key}
\item{Cas}{Cas system found in the whole genome, Cas_id is the key, one Cas_id contains multi cas protein}
\item{Arrary}{the arrary structure of each CRISPR, "LeftFLANK","CRISPRdr","CRISPRspacer","RightFLANK"}
\item{Spacer}{all spacer come from each CRISPR, Spacer_id is the key}
}
}
\usage{
crispr
}
\description{
a list contains four dataframe.
}
