% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{gene_pd_lm}
\alias{gene_pd_lm}
\title{Lm for gene_not_gene ~ phy_dis}
\usage{
gene_pd_lm(arc_count, mode = 1, filter_spacer = 5, point_param = list())
}
\arguments{
\item{arc_count}{data.frame which columns: "source_name","target_name","spacer_n","phy_dis","gene_ratio".}

\item{mode}{1~2}

\item{filter_spacer}{filter_spacer}

\item{point_param}{point_param}
}
\value{
ggplot
}
\description{
Lm for gene_not_gene ~ phy_dis
}
