% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_structrue.R
\name{get_Arrays}
\alias{get_Arrays}
\title{Get Arrays from crispr.}
\usage{
get_Arrays(crispr)
}
\arguments{
\item{crispr}{crispr object}
}
\value{
Arrays
}
\description{
Get Arrays from crispr.
}
\examples{
data(crispr)
Arrays=get_Arrays(crispr)
}
