% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{get_consensus_rep_fa}
\alias{get_consensus_rep_fa}
\title{Get consensus repeat fasta}
\usage{
get_consensus_rep_fa(crispr, evidence_level = NULL, cas = NULL)
}
\arguments{
\item{crispr}{crispr}

\item{evidence_level}{filter the evidence_level (1~4), if NULL, then no filter}

\item{cas}{filter with cas system or not (TRUE or FALSE), if NULL, then no filter}
}
\value{
fasta
}
\description{
Get consensus repeat fasta
}
\examples{
data(crispr)
get_consensus_rep_fa(crispr)
}
