% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{get_spacer_fa}
\alias{get_spacer_fa}
\title{Get spacer fasta}
\usage{
get_spacer_fa(crispr, evidence_level = NULL, cas = NULL)
}
\arguments{
\item{crispr}{crispr}

\item{evidence_level}{filter the evidence_level (1~4), if NULL, then no filter}

\item{cas}{filter with cas system or not (T or F), if NULL, then no filter}
}
\description{
Get spacer fasta
}
\examples{
data(crispr)
get_spacer_fa(crispr,evidence_level=4,cas=TRUE)
}
