% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCRISPR-package.R
\docType{package}
\name{iCRISPR-package}
\alias{iCRISPR}
\alias{iCRISPR-package}
\title{iCRISPR: CRISPR Analysis}
\description{
CRISPR analysis including CRISPR identification, prepare, summary and visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Asa12138/iCRISPR}
  \item Report bugs at \url{https://github.com/Asa12138/iCRISPR/issues}
}

}
\author{
\strong{Maintainer}: Chen Peng \email{pengchen2001@zju.edu.cn} (\href{https://orcid.org/0000-0002-9449-7606}{ORCID})

Authors:
\itemize{
  \item Zhen Liu \email{lz19@zju.edu.cn}
}

}
\keyword{internal}
