% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_structrue.R
\name{install_mafft}
\alias{install_mafft}
\title{Install MAFFT}
\usage{
install_mafft(mafft.zip = NULL, install_dir = tools::R_user_dir("iCRISPR"))
}
\arguments{
\item{mafft.zip}{your download mafft.zip file from https://mafft.cbrc.jp/alignment/software/}

\item{install_dir}{The directory to install MAFFT.}
}
\value{
No value
}
\description{
This function downloads and installs MAFFT at the specified path.
}
