% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_structrue.R
\name{is_same_seq}
\alias{is_same_seq}
\title{Judge two sequence are same or not}
\usage{
is_same_seq(s1, s2, xmode = FALSE)
}
\arguments{
\item{s1}{First sequence}

\item{s2}{Second sequence}

\item{xmode}{consider the reverse string (all), complement and reverse complement string (DNA)}
}
\value{
Sequence identity as a decimal value between 0 and 1
}
\description{
Judge two sequence are same or not
}
\examples{
is_same_seq("ATCGTACG","ATCGTACG")
is_same_seq("ATCGTACG","ATCGTACG",xmode=TRUE)
}
