% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{multi_pre_CCF_res}
\alias{multi_pre_CCF_res}
\title{Prepare many genome result files from CRISPRCasFinder}
\usage{
multi_pre_CCF_res(
  input_folder,
  output_folder = NULL,
  threads = 1,
  verbose = TRUE
)
}
\arguments{
\item{input_folder}{the folder of CRISPRCasFinder result, contains GFF/,TSV/,result.json,...}

\item{output_folder}{output folder, default: NULL, don not output to file.}

\item{threads}{threads, default: 1}

\item{verbose}{verbose}
}
\value{
No value
}
\description{
Prepare many genome result files from CRISPRCasFinder
}
