% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{plot.self_df}
\alias{plot.self_df}
\title{plot self_df}
\usage{
\method{plot}{self_df}(x, mode = 1, filter_spacer = 0, ...)
}
\arguments{
\item{x}{self_df}

\item{mode}{1~2}

\item{filter_spacer}{filter spacer less than filter_spacer, default 0}

\item{...}{add}
}
\value{
ggplot
}
\description{
plot self_df
}
