% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{plot.tri_target_tree}
\alias{plot.tri_target_tree}
\title{Plot tri_target_tree}
\usage{
\method{plot}{tri_target_tree}(
  x,
  add_self = TRUE,
  add_virus = TRUE,
  add_others = TRUE,
  add_tiplab = TRUE,
  some_tax = c(),
  label_size = 1.5,
  ...
)
}
\arguments{
\item{x}{tri_target_tree}

\item{add_self}{logical}

\item{add_virus}{logical}

\item{add_others}{logical}

\item{add_tiplab}{logical}

\item{some_tax}{vectors}

\item{label_size}{label size}

\item{...}{add}
}
\value{
ggplot
}
\description{
Plot tri_target_tree
}
