% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{plot_cas_type}
\alias{plot_cas_type}
\title{Plot the Cas system type and subtype}
\usage{
plot_cas_type(crispr)
}
\arguments{
\item{crispr}{crispr result from \code{pre_CCF_res()}}
}
\value{
ggplot
}
\description{
Plot the Cas system type and subtype
}
\examples{
plot_cas_type(crispr)
}
