% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{pre_CCF_res}
\alias{pre_CCF_res}
\title{Prepare the result files from CRISPRCasFinder}
\usage{
pre_CCF_res(
  input_folder,
  output_folder = "./pre_CCF_res_out",
  genome_name = NULL,
  verbose = T
)
}
\arguments{
\item{input_folder}{the folder of CRISPRCasFinder result, contains GFF/,TSV/,result.json,...}

\item{output_folder}{output, default: ./}

\item{genome_name}{the genome_name}

\item{verbose}{verbose, default: T}
}
\description{
Prepare the result files from CRISPRCasFinder
}
\examples{
MAG_test=system.file("extdata/MAG_test",package = "iCRISPR")
crispr=pre_CCF_res(MAG_test,output_folder="./pre_CCF_res_out")
}
\references{
https://crisprcas.i2bc.paris-saclay.fr/CrisprCasFinder/Index
}
