% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{print.crispr}
\alias{print.crispr}
\title{Print crispr object}
\usage{
\method{print}{crispr}(x, ...)
}
\arguments{
\item{x}{crispr object}

\item{...}{add}
}
\description{
Print crispr object
}
\examples{
data(crispr)
print(crispr)
}
