% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.multi_crispr}
\alias{print.multi_crispr}
\title{Print multi_crispr object}
\usage{
\method{print}{multi_crispr}(x, ...)
}
\arguments{
\item{x}{multi_crispr object}

\item{...}{add}
}
\value{
No value
}
\description{
Print multi_crispr object
}
