% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seqinfo.R
\name{random_seq}
\alias{random_seq}
\title{Generate Random Sequences}
\usage{
random_seq(
  num_seqs = 100,
  mean_length = 32,
  sd_length = 3,
  element = c("A", "T", "C", "G")
)
}
\arguments{
\item{num_seqs}{The number of sequences to generate.}

\item{mean_length}{The mean length of the sequences.}

\item{sd_length}{The standard deviation of the sequence lengths.}

\item{element}{default: c("A", "T", "C", "G")}
}
\value{
A data frame with sequence names and sequences.
}
\description{
Generate Random Sequences
}
\examples{
random_seq()

}
