% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seqinfo.R
\name{random_tandem_repeat}
\alias{random_tandem_repeat}
\title{Generate Random Tandem Repeat Sequences}
\usage{
random_tandem_repeat(
  num_seqs = 100,
  mean_unit_length = 6,
  sd_unit_length = 1,
  mean_num_units = 5,
  sd_num_units = 2,
  element = c("A", "T", "C", "G")
)
}
\arguments{
\item{num_seqs}{The number of sequences to generate.}

\item{mean_unit_length}{The mean length of the repeat unit.}

\item{sd_unit_length}{The standard deviation of the repeat unit lengths.}

\item{mean_num_units}{The mean number of repeat units in each sequence.}

\item{sd_num_units}{The standard deviation of the number of repeat units.}

\item{element}{Default: c("A", "T", "C", "G")}
}
\value{
A data frame with sequence names and tandem repeat sequences.
}
\description{
Generate Random Tandem Repeat Sequences
}
\examples{
random_tandem_repeat()
}
