% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{select_crispr}
\alias{select_crispr}
\title{Select part of multi_crispr.}
\usage{
select_crispr(.data, ...)
}
\arguments{
\item{.data}{multi_crispr}

\item{...}{additional}
}
\value{
multi_crispr
}
\description{
Select part of multi_crispr.
}
\examples{
data(multi_crispr)
select_crispr(multi_crispr,1:10)
}
