% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCRISPR-package.R
\name{set_iCRISPR_config}
\alias{set_iCRISPR_config}
\title{Set config}
\usage{
set_iCRISPR_config(item, value)
}
\arguments{
\item{item}{item}

\item{value}{value}
}
\value{
No value
}
\description{
Set config
}
