% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_type.R
\name{show_cas_type}
\alias{show_cas_type}
\title{Show cas type}
\usage{
show_cas_type()
}
\value{
ggplot
}
\description{
Show cas type
}
\examples{
show_cas_type()
}
\references{
\enumerate{
\item Makarova, K. S. et al. Evolutionary classification of CRISPR-Cas systems: a burst of class 2 and derived variants. Nat Rev Microbiol 18, 67–83 (2020).
}
}
