% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{source_target_count}
\alias{source_target_count}
\title{Summary the phylogenetic distance and hit count between source and target}
\usage{
source_target_count(source_target_info)
}
\arguments{
\item{source_target_info}{data.frame which columns: "source_name","source_lineage","target_name","target_lineage"}
}
\value{
s_t_count
}
\description{
Summary the phylogenetic distance and hit count between source and target
}
\examples{
data(pro_net)
source_target_count(pro_net)
}
