% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{spacer_pd_lm}
\alias{spacer_pd_lm}
\title{Lm for spacer_n ~ phy_dis}
\usage{
spacer_pd_lm(
  arc_count,
  mode = 1,
  filter_spacer = 5,
  target_tax = NULL,
  remove_viruses = TRUE
)
}
\arguments{
\item{arc_count}{data.frame which columns: "source_name","target_name","spacer_n","phy_dis".}

\item{mode}{1~4}

\item{filter_spacer}{filter_spacer}

\item{target_tax}{NULL or target_tax}

\item{remove_viruses}{TRUE for mode 3~4}
}
\value{
ggplot
}
\description{
Lm for spacer_n ~ phy_dis
}
