% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_type.R
\name{summary_cas_type}
\alias{summary_cas_type}
\alias{plot.cas_type}
\title{Summary Cas-system type}
\usage{
summary_cas_type(crispr)

\method{plot}{cas_type}(x, mode = 1, ...)
}
\arguments{
\item{crispr}{crispr object}

\item{x}{cas_type object}

\item{mode}{1~3}

\item{...}{additional}
}
\value{
ggplot
}
\description{
Summary Cas-system type

Plot the Cas system type and subtype
}
\examples{
data(multi_crispr)
cas_type_res=summary_cas_type(multi_crispr)
plot(cas_type_res)
}
