% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{summary_levels}
\alias{summary_levels}
\title{Summary crispr and spacer number of differnent evidence levels}
\usage{
summary_levels(crispr)
}
\arguments{
\item{crispr}{crispr object}
}
\description{
Summary crispr and spacer number of differnent evidence levels
}
\examples{
summary_levels(crispr)
}
