% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{tri_target}
\alias{tri_target}
\title{Summary spacers target which: Self? Virus? Others?}
\usage{
tri_target(source_target_info, two_level = NULL)
}
\arguments{
\item{source_target_info}{data.frame which columns: "source_lineage","target_lineage"}

\item{two_level}{manual which two levels in "Kingdom","Phylum","Class","Order","Family","Genus","Species". Sometimes "Genome"}
}
\value{
data.frame
}
\description{
Summary spacers target which: Self? Virus? Others?
}
\examples{
tri_target(pro_net, two_level = c("Species", "Genus"))
}
