% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{tri_target_tree}
\alias{tri_target_tree}
\title{Build tri_target_tree}
\usage{
tri_target_tree(source_target_info, self_df, level = 7, filter_spacer = 0)
}
\arguments{
\item{source_target_info}{data.frame which columns: "source_lineage","target_lineage"}

\item{self_df}{self_df from \code{tri_target}}

\item{level}{1~7, means "Kingdom","Phylum","Class","Order","Family","Genus","Species"}

\item{filter_spacer}{filter spacer less than filter_spacer, default 0}
}
\value{
tree
}
\description{
Build tri_target_tree
}
\examples{
self_df <- tri_target(pro_net)
tree <- tri_target_tree(pro_net, self_df, filter_spacer = 10)
# plot(tree)
}
