% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_type.R
\name{combine.cas_type}
\alias{combine.cas_type}
\title{Combine some cas_type object}
\usage{
\method{combine}{cas_type}(..., each_genome = T)
}
\arguments{
\item{...}{cas_type object}
}
\value{
cas_type object
}
\description{
Combine some cas_type object
}
