% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_structrue.R
\name{compare_array}
\alias{compare_array}
\alias{plot.array_comparison}
\alias{align_array}
\title{Compare two crispr arrays (whatever arrays indeed)}
\usage{
compare_array(array1, array2)

\method{plot}{array_comparison}(x, linewidth = 2, ...)

align_array(array_comparison)
}
\arguments{
\item{array1}{a vector contains spacers (with order) as reference.}

\item{array2}{a vector contains spacers (with order)}

\item{x}{array_comparison}

\item{...}{add}

\item{array_comparison}{array_comparison object}
}
\value{
array_comparison object

ggplot

aligned array_comparison object
}
\description{
Compare two crispr arrays (whatever arrays indeed)

Plot array_comparison

Align array_comparison
}
\examples{
array_test=random_seq(5)[,2]
res=compare_array(array1=array_test,array2=array_test[c(5,1:3)])
plot.array_comparison(res)
align_array(res)->align_res
plot.array_comparison(align_res)
}
