% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iCRISPR-package.R
\docType{package}
\name{iCRISPR-package}
\alias{iCRISPR}
\alias{iCRISPR-package}
\title{iCRISPR: CRISPR analysis}
\description{
CRISPR analysis including CRISPR identification, prepare, summary and visualization.
}
\author{
\strong{Maintainer}: Chen Peng \email{pengchen2001@zju.edu.cn}

Authors:
\itemize{
  \item Zhen Liu \email{lz19@zju.edu.cn}
}

}
\keyword{internal}
