% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{multi_pre_CCF_res}
\alias{multi_pre_CCF_res}
\title{Prepare many genome result files from CRISPRCasFinder}
\usage{
multi_pre_CCF_res(
  input_folder,
  output_folder = "./pre_CCF_res_out",
  threads = 1
)
}
\arguments{
\item{input_folder}{the folder of CRISPRCasFinder result, contains GFF/,TSV/,result.json,...}

\item{output_folder}{output, default: ./}

\item{threads}{threads, default:4}
}
\description{
Prepare many genome result files from CRISPRCasFinder
}
\examples{
multi_res=multi_pre_CCF_res("extdata",threads=1)
}
