% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{pre_CCF_res}
\alias{pre_CCF_res}
\title{Prepare the result files from CRISPRCasFinder}
\usage{
pre_CCF_res(input_folder, output_folder = "./", genome_name = NULL)
}
\arguments{
\item{input_folder}{the folder of CRISPRCasFinder result, contains GFF/,TSV/,result.json,...}

\item{output_folder}{output, default: ./}

\item{genome_name}{the genome_name}
}
\description{
Prepare the result files from CRISPRCasFinder
}
\examples{
pre_CCF_res("extdata/AAB-S01R1_122","extdata")
}
