% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRISPRCasfinder_downstream.R
\name{print.multi_crispr}
\alias{print.multi_crispr}
\title{Print multi_crispr object}
\usage{
\method{print}{multi_crispr}(multi_crispr, ...)
}
\arguments{
\item{multi_crispr}{multi_crispr object}

\item{...}{add}
}
\description{
Print multi_crispr object
}
