% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tri_target.R
\name{sankey_overview}
\alias{sankey_overview}
\title{Sankey plot for source-target}
\usage{
sankey_overview(
  arc_count,
  source_tax = NULL,
  target_tax = NULL,
  two_level = NULL,
  topN1 = 8,
  topN2 = 8,
  file = NULL
)
}
\arguments{
\item{arc_count}{data.frame which columns: "source_name","target_name","spacer_n".}

\item{source_tax}{source_tax table}

\item{target_tax}{target_tax table}

\item{two_level}{manual which two levels in "Kingdom","Phylum","Class","Order","Family","Genus","Species". Sometimes "Genome"}

\item{topN1}{select topN for level1}

\item{topN2}{select topN for level2}

\item{file}{filename}
}
\value{
ggplot
}
\description{
Sankey plot for source-target
}
\examples{
data(pro_net)
sankey_overview(pro_net, two_level = c("Species", "Genus"))
}
