% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cas_type.R
\name{summary_levels}
\alias{summary_levels}
\alias{plot.evidence_level}
\title{Summary crispr and spacer number of different evidence levels}
\usage{
summary_levels(crispr, each_genome = T)

\method{plot}{evidence_level}(x, mode = 1, ...)
}
\arguments{
\item{crispr}{crispr object}

\item{each_genome}{summary each genome information in multi_crispr?}

\item{x}{evidence_level object}

\item{mode}{1~2}

\item{...}{additional}
}
\value{
ggplot
}
\description{
Summary crispr and spacer number of different evidence levels

Plot the distribution of array and spacer number at different evidence levels.
}
\examples{
level_res=summary_levels(multi_crispr,each_genome=FALSE)
plot(level_res)
}
