% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seqinfo.R
\name{summary_seq}
\alias{summary_seq}
\title{Summarize sequence information}
\usage{
summary_seq(sequence, max_k = 7, seq_col = 2, threads = 1)
}
\arguments{
\item{sequence}{The input DNA sequence.}

\item{max_k}{The maximum k value for kmers.}

\item{seq_col}{the column name of sequence if your input is a dataframe.}

\item{threads}{threads}
}
\description{
Summarize sequence information
}
\examples{
summary_seq("CCTGAACCTATGCCGTCCACCTTGCGTTGCCTT")
summary_seq(random_seq(10))
}
